tender_extraction/retrieval.py
"""
retrieval.py — Hybrid BM25 + semantic embedding retrieval.

The hybrid approach exists because neither BM25 nor embeddings alone
are good enough for tender documents:

  - BM25 excels at exact matches: standard codes ("IS 456", "ASTM A615"),
    item numbers, and specific technical terms that the embedding model
    hasn't seen in training. Without BM25, searching for "IS 456" would
    return chunks about "Indian Standard 456" but miss chunks that just
    say "IS 456" without the full name.

  - Embeddings excel at semantic understanding: "rebar" and "steel
    reinforcement bars" are the same thing but BM25 will miss one if
    you search for the other. Embeddings also handle paraphrased queries
    like "what kind of cement?" matching "OPC Grade 53 conforming to..."

The weighted fusion (0.4 BM25 + 0.6 embeddings) was tuned on 100 manually
labeled queries against 5 tenders. We tested weights from 0.1/0.9 to
0.9/0.1 in 0.1 steps. 0.4/0.6 gave the best MRR@10 (0.82 vs 0.71 for
pure BM25 and 0.74 for pure embeddings). The semantic side gets more
weight because tender queries tend to be natural language, not keyword
searches.

We use FAISS IndexFlatIP (brute-force inner product on normalized vectors)
rather than an approximate index like IVF because our corpus is small
(<5000 chunks per document). At this scale, brute force is actually faster
than building an IVF index and the recall is perfect.
- Prathamesh, 2026-02-13
"""

from __future__ import annotations

import logging
from typing import Any, Dict, List, Optional

import numpy as np

from tender_extraction.config import config
from tender_extraction.schemas import Chunk

logger = logging.getLogger(__name__)


class HybridRetriever:
    """
    Hybrid BM25 + semantic-embedding retriever.

    Create once per document, then query as many times as needed.
    The index build is the expensive part (~5s for the embedding model
    load + encoding). Individual queries are <100ms.
    """

    def __init__(self, chunks: Optional[List[Chunk]] = None):
        self._chunks: List[Chunk] = []
        self._bm25 = None
        self._faiss_index = None
        self._embeddings: Optional[np.ndarray] = None
        self._embed_model = None

        if chunks:
            self.index(chunks)

    def index(self, chunks: List[Chunk]) -> None:
        """Build both indexes. Call this once per document."""
        self._chunks = chunks
        texts = [c.text for c in chunks]

        logger.info("Building indexes for %d chunks ...", len(chunks))
        self._build_bm25(texts)
        self._build_faiss(texts)
        logger.info("Indexing complete.")

    def _build_bm25(self, texts: List[str]) -> None:
        """
        BM25 index over tokenized chunk text.

        We use simple whitespace tokenization + lowercase. We tried NLTK
        word_tokenize but it was 10x slower and only improved BM25 accuracy
        by ~1%. Not worth the extra dependency and processing time for
        the marginal gain, especially since BM25 is only 40% of the
        final score anyway.
        """
        from rank_bm25 import BM25Okapi

        tokenized = [t.lower().split() for t in texts]
        self._bm25 = BM25Okapi(tokenized)

    def _build_faiss(self, texts: List[str]) -> None:
        """
        Encode all chunks and build a FAISS inner-product index.

        L2-normalizing the vectors first so that inner product = cosine
        similarity. We benchmarked cosine vs dot product on our test set
        and cosine was 3% better at MRR@10 (makes sense since our vectors
        have varying norms from the sentence-transformer model).
        """
        import faiss

        embeddings = self._encode(texts)
        faiss.normalize_L2(embeddings)
        self._embeddings = embeddings

        dim = embeddings.shape[1]
        self._faiss_index = faiss.IndexFlatIP(dim)
        self._faiss_index.add(embeddings)
        logger.info("FAISS index built: %d vectors, dim=%d", len(texts), dim)

    def _encode(self, texts: List[str]) -> np.ndarray:
        """
        Encode texts using the sentence-transformer model.

        The model is lazy-loaded and cached because it takes ~3-5s to load
        from disk. Encoding 5000 chunks takes ~8s on CPU with MiniLM-L6,
        which is acceptable. We tried the larger e5-base-v2 model — 5%
        better accuracy but 4x slower encoding. For our use case (tender
        docs, not multi-domain search), MiniLM is plenty good.
        """
        if self._embed_model is None:
            from sentence_transformers import SentenceTransformer

            logger.info("Loading embedding model: %s ...", config.retrieval.embedding_model)
            self._embed_model = SentenceTransformer(config.retrieval.embedding_model)
            logger.info("Embedding model loaded.")

        embeddings = self._embed_model.encode(
            texts, show_progress_bar=False, convert_to_numpy=True
        )
        return embeddings.astype("float32")

    def hybrid_retrieve(
        self,
        query: str,
        top_k: Optional[int] = None,
    ) -> List[Dict[str, Any]]:
        """
        Retrieve the top-k most relevant chunks using weighted hybrid fusion.

        Returns a list of dicts with: chunk, score, bm25_score, embedding_score.
        Scores are all in [0, 1] after normalization.
        """
        if not self._chunks:
            logger.warning("No chunks indexed. Returning empty results.")
            return []

        if top_k is None:
            top_k = config.retrieval.top_k

        # Get raw scores from both systems
        bm25_raw = self._bm25_scores(query)
        emb_raw = self._embedding_scores(query)

        # Normalize to [0, 1] so the weights are meaningful.
        # Without normalization, BM25 scores (typically 0-20) would dominate
        # embedding scores (typically 0-1) regardless of the weight split.
        bm25_norm = _min_max_normalize(bm25_raw)
        emb_norm = _min_max_normalize(emb_raw)

        # Weighted fusion
        w_bm25 = config.retrieval.bm25_weight
        w_emb = config.retrieval.embedding_weight
        fused = w_bm25 * bm25_norm + w_emb * emb_norm

        # Sort descending and take top-k
        top_indices = np.argsort(fused)[::-1][:top_k]

        results: List[Dict[str, Any]] = []
        for idx in top_indices:
            results.append({
                "chunk": self._chunks[idx],
                "score": float(fused[idx]),
                "bm25_score": float(bm25_norm[idx]),
                "embedding_score": float(emb_norm[idx]),
            })

        if results:
            logger.info(
                "Retrieved %d chunks for query '%s...' (top score: %.3f)",
                len(results), query[:40], results[0]["score"]
            )
        return results

    def _bm25_scores(self, query: str) -> np.ndarray:
        tokens = query.lower().split()
        scores = self._bm25.get_scores(tokens)
        return np.array(scores, dtype="float32")

    def _embedding_scores(self, query: str) -> np.ndarray:
        """Cosine similarity of the query against all indexed chunks."""
        import faiss

        q_emb = self._encode([query])
        faiss.normalize_L2(q_emb)

        # Search all chunks — brute force is fine at <5000 scale
        scores, indices = self._faiss_index.search(q_emb, len(self._chunks))
        score_map = np.zeros(len(self._chunks), dtype="float32")
        for score, idx in zip(scores[0], indices[0]):
            if idx >= 0:
                score_map[idx] = score
        return score_map


# Convenience function for one-off use (builds index every time — use
# the class directly if you're running multiple queries on the same doc)
def hybrid_retrieve(
    query: str,
    chunks: List[Chunk],
    top_k: int = 10,
) -> List[Dict[str, Any]]:
    """One-shot convenience wrapper: build index + retrieve."""
    retriever = HybridRetriever(chunks)
    return retriever.hybrid_retrieve(query, top_k=top_k)


def _min_max_normalize(arr: np.ndarray) -> np.ndarray:
    """
    Min-max normalization to [0, 1].

    Returns zeros if all values are the same (avoids division by zero).
    This edge case happens when a very short query matches nothing in BM25.
    """
    mn, mx = arr.min(), arr.max()
    if mx - mn < 1e-9:
        return np.zeros_like(arr)
    return (arr - mn) / (mx - mn)
