requirements.txt
# TenderExtractPro - Dependencies
# Install: pip install -r requirements.txt

# PDF & Document Processing
pdfplumber>=0.10.0
python-docx>=1.1.0
pdf2image>=1.16.3

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Retrieval & Embeddings
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# LLM Inference
llama-cpp-python>=0.2.0

# Validation & Schemas
pydantic>=2.0.0

# Token Counting
tiktoken>=0.5.0

# Core
numpy>=1.24.0
